+++ src/blocks.js
@@ -161,7 +161,7 @@ CostumeIconMorph, SoundIconMorph, SVG_Costume*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.blocks = '2022-March-15';
+modules.blocks = '2022-March-22';
 
 var SyntaxElementMorph;
 var BlockMorph;
@@ -3797,7 +3797,8 @@ BlockMorph.prototype.restoreInputs = function (oldInputs, offset = 0) {
     else if (oldInputs.length &&
         (inputs.length === 1) &&
         trg instanceof MultiArgMorph &&
-        !(src instanceof MultiArgMorph)
+        !(src instanceof MultiArgMorph) &&
+        !(src instanceof ArgLabelMorph)
     ) {
         element = trg;
         inputs = element.inputs();
@@ -3987,12 +3988,15 @@ BlockMorph.prototype.exportScript = function () {
     }
 };
 
-BlockMorph.prototype.toXMLString = function () {
+BlockMorph.prototype.toXMLString = function (receiver) {
+    // answer an xml string representation of this block and all the
+    // following ones attached to it, including all dependencies.
+    // specifying a receiver sprite is optional for cases where
+    // the receiver sprite is not the currently edited one inside the IDE
     var ide = this.parentThatIsA(IDE_Morph),
         blockEditor = this.parentThatIsA(BlockEditorMorph),
-        rcvr = this.scriptTarget(),
-        dependencies = [],
-        isReporter = this instanceof ReporterBlockMorph;
+        isReporter = this instanceof ReporterBlockMorph,
+        dependencies;
 
     if (!ide && blockEditor) {
         ide = blockEditor.target.parentThatIsA(IDE_Morph);
@@ -4002,31 +4006,47 @@ BlockMorph.prototype.toXMLString = function () {
     }
 
     // collect custom block definitions referenced in this script:
-    this.forAllChildren(morph => {
-        var def;
-        if (morph.isCustomBlock) {
-            def = morph.isGlobal ? morph.definition
-                : rcvr.getMethod(morph.semanticSpec);
-            [def].concat(def.collectDependencies([], [], rcvr)).forEach(
-                fun => {
-                    if (!contains(dependencies, fun)) {
-                        dependencies.push(fun);
-                    }
-                }
-            );
-        }
-    });
+    dependencies = this.dependencies(false, receiver); // both global and local
 
     return '<script app="' +
         ide.serializer.app +
         '" version="' +
         ide.serializer.version +
         '">' +
-        (dependencies.length ? ide.blocksLibraryXML(dependencies, false) : '') +
+        (dependencies.length ? ide.blocksLibraryXML(dependencies) : '') +
         (isReporter ? '<script>' : '') +
         ide.serializer.serialize(this) +
         (isReporter ? '</script>' : '') +
         '</script>';
+};
+
+BlockMorph.prototype.dependencies = function (onlyGlobal, receiver) {
+    // answer an array containing all custom block definitions referenced
+    // by this and the following blocks, optional parameter to constrain
+    // to global definitions.
+    // specifying a receiver sprite is optional for cases where
+    // the receiver sprite is not the currently edited one inside the IDE
+    // if a receiver is not specified  this method can only be called from
+    // within the IDE because it needs to be able to determine the scriptTarget
+    var dependencies = [],
+        rcvr = onlyGlobal ? null : (receiver || this.scriptTarget());
+    this.forAllChildren(morph => {
+        var def;
+        if (morph.isCustomBlock) {
+            if (!onlyGlobal || (onlyGlobal && morph.isGlobal)) {
+                def = morph.isGlobal ? morph.definition
+                    : rcvr.getMethod(morph.semanticSpec);
+                [def].concat(def.collectDependencies([], [], rcvr)).forEach(
+                    fun => {
+                        if (!contains(dependencies, fun)) {
+                            dependencies.push(fun);
+                        }
+                    }
+                );
+            }
+        }
+    });
+    return dependencies;
 };
 
 // BlockMorph syntax analysis